+++ src/main/java/de/alpharogroup/collections/list/ListExtensions.java
@@ -480,6 +480,26 @@ public final class ListExtensions
 	}
 
 	/**
+	 * To array.
+	 *
+	 * @param <T>
+	 *            the generic type
+	 * @param list
+	 *            the list
+	 * @return the array or null if the list is empty
+	 */
+	@SuppressWarnings("unchecked")
+	public static <T> T[] toArray(final @NonNull List<T> list)
+	{
+		if (list.isEmpty())
+		{
+			throw new IllegalArgumentException("list is empty");
+		}
+		T[] newArray = (T[])ArrayFactory.newArray(list.get(0).getClass(), list.size());
+		return list.toArray(newArray);
+	}
+
+	/**
 	 * Converts the given enumaration to a Vector.
 	 *
 	 * @param <T>